requirements.txt
Flask~=2.2.3
requests~=2.31.0
pandas~=1.5.3
pyotp~=2.9.0
qrcode~=7.4.2
plotly~=5.18.0
django-qrcode~=0.3
python-dotenv~=1.0.1